refactor: 컴포넌트 분리 및 타입 공통화
